// EHS_FrontEnd_Angular/src/app/interfaces/UnitMaster.class.ts
export class UnitMaster {
    unitId: string;
    unitCode: string;
    unitDisplayName: string;
    divisionId: string;
    divisionName: string;
    unitUserName: string;
    unitUserPassword: string;
    unitEHSHead: string;
    unitEHSHeadUser: string;
    unitEHSHeadPhone: string;
    unitEHSAdmin: string;
    unitEHSAdminUser: string;
    unitEHSAdminPhone: string;
    unitHead: string;
    unitHeadUser: string;
    unitHeadPhone: string;
    unitStatus: string;
    createdUTCDate: string;
    createdBy: string;
    modifiedBy: string;
    modifiedUTCDate: string;
    modules: string;
}
export class UnitMasterAddorUpdate {
    UnitId: string;
    UnitCode: string;
    unitDisplayName: string;
    UnitName: string;
    DivisionId: string;
    UnitUserName: string;
    UnitUserPassword: string;
    UnitEHSHead: string;
    UnitEHSAdmin: string;
    UnitHead: string;
    UnitStatus: boolean;
    CreatedBy: string;
    ModifiedBy: string;
    Modoules: string;
}